requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
drf-yasg
pillow
celery
redis
djangorestframework-simplejwt
gunicorn
django-filter
django-cors-headers
django-autoslug
